fix: replace nested Stack with Slot in recipes layout to fix Android ScreenStackFragment crash
The recipes layout used a <Stack> navigator inside the root <Stack>, creating
nested ScreenStack fragments on Android. This caused the 'ScreenStackFragment
added into a non-stack container' crash when navigating to recipe/food editors.

Replace the inner Stack with <Slot> and move header configuration into each
screen component via <Stack.Screen> options.

Resolves #94
